+++ packages/sdk/src/api-functions/walletAuth/index.ts
@@ -1,0 +1,3 @@
+export * from './requestWalletChallenge';
+export * from './verifyWalletSignature';
+export * from './revokeWalletToken';